fix if image not exist, image URL

+++ src/java/controllers/LoadImageServlet.java
@@ -34,10 +34,13 @@ public class LoadImageServlet extends HttpServlet {
             throws ServletException, IOException {
         String filename = URLDecoder.decode(request.getPathInfo().substring(1), "UTF-8");
         File file = new File("/Users/tuannnh/Desktop/images", filename);
-        response.setHeader("Content-Type", getServletContext().getMimeType(filename));
-        response.setHeader("Content-Length", String.valueOf(file.length()));
-        response.setHeader("Content-Disposition", "inline; filename=\"" + file.getName() + "\"");
-        Files.copy(file.toPath(), response.getOutputStream());
+        if (file.exists()) {
+            response.setHeader("Content-Type", getServletContext().getMimeType(filename));
+            response.setHeader("Content-Length", String.valueOf(file.length()));
+            response.setHeader("Content-Disposition", "inline; filename=\"" + file.getName() + "\"");
+            Files.copy(file.toPath(), response.getOutputStream());
+        }
+
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
